Organize routes by file
